+++ ArtSpace/src/main/java/com/example/demo/user/service/CorporationService.java
@@ -1,0 +1,5 @@
+package com.example.demo.user.service;
+
+public interface CorporationService {
+
+}
