+++ ArtSpace/src/main/java/com/example/demo/user/service/CorporationServiceImpl.java
@@ -1,0 +1,8 @@
+package com.example.demo.user.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CorporationServiceImpl {
+
+}
